+++ app/Http/Controllers/StoryController.php
@@ -130,83 +130,4 @@ class StoryController extends Controller
 
         return redirect()->route('listStoryPage');
     }
-
-    public function listChapterPage($slug)
-    {
-        $story  = Story::where('slug', $slug)->get();
-        $listChapter = Chapter::where('story_id', $story[0]->id)
-            ->orderBy('created_at', 'desc')
-            ->get();
-
-        return view('chapter.index', [
-            'slug' => $slug,
-            'listChapter' => $listChapter
-        ]);
-    }
-
-    public function createChapterPage($slug)
-    {
-        $story = Story::where('slug', $slug)
-            ->get();
-
-        return view('chapter.create', [
-            'slug' => $slug,
-            'story_id' => $story[0]->id
-        ]);
-    }
-
-    public function storeChapter(Request $request, $slug)
-    {
-        $request->validate([
-            'story_id' => 'required|integer',
-            'name' => 'required',
-            'content' => 'required',
-        ]);
-
-        $chapterNumber = Chapter::where('story_id', $request->story_id)->get()->max('chapter_number');
-
-        if ($chapterNumber) {
-            $chapterNumber++;
-        } else {
-            $chapterNumber = 1;
-        }
-
-        $chapter = Chapter::create([
-            'slug' => Str::slug($request->name, '_') . '_' . time(),
-            'chapter_number' => $chapterNumber,
-            'name' => $request->name,
-            'content' => $request->content,
-            'story_id' => $request->story_id,
-        ]);
-
-        return redirect()->route('listChapterPage', ['slug' => $slug]);
-    }
-
-    public function editChapterPage($slug_story, $slug_chapter)
-    {
-        $chapter = Chapter::where('slug', $slug_chapter)->first();
-
-        return view('chapter.edit', ['slug_story' => $slug_story, 'chapter' => $chapter]);
-    }
-
-    public function updateChapter(Request $request, $slug_story, $slug_chapter)
-    {
-        $request->validate([
-            'name' => 'required',
-            'content' => 'required',
-        ]);
-
-        $chapter = Chapter::where('slug', $slug_chapter)->first();
-        $chapter->name = $request->input('name');
-        $chapter->content  = $request->input('content');
-        $chapter->save();
-        return redirect()->route('listChapterPage', ['slug' => $slug_story]);
-    }
-
-    public function deleteChapter($slug_story, $slug_chapter)
-    {
-        $chapter = Chapter::where('slug', $slug_chapter)->first();
-        $chapter->delete();
-        return redirect()->route('listChapterPage', ['slug' => $slug_story]);
-    }
 }
