+++ app/Providers/AuthServiceProvider.php
@@ -4,8 +4,10 @@ namespace App\Providers;
 
 // use Illuminate\Support\Facades\Gate;
 
+use App\Models\Chapter;
 use App\Models\Story;
 use App\Models\User;
+use App\Policies\ChapterPolicy;
 use App\Policies\StoryPolicy;
 use App\Policies\UserPolicy;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
@@ -19,7 +21,8 @@ class AuthServiceProvider extends ServiceProvider
      */
     protected $policies = [
         User::class => UserPolicy::class,
-        Story::class => StoryPolicy::class
+        Story::class => StoryPolicy::class,
+        Chapter::class => ChapterPolicy::class
     ];
 
     /**
